job_sequence_generator: code cleanup

+++ code/events/headers/Event.h
@@ -2,57 +2,22 @@
 // Created by mihael on 29/04/23.
 //
 
-/**
- * @file Event.h
- * @brief Defines the Event class for representing an event in the scheduling simulation.
- */
-
 #ifndef SCHEDULING_EVENT_H
 #define SCHEDULING_EVENT_H
 
 #include "EventType.h"
 #include "string"
 
-/**
- * @class Event
- * @brief Represents an event in the scheduling simulation.
- *
- * The Event class encapsulates information about an event that occurs in the scheduling simulation process.
- * It includes attributes to store the event time, event type, and an optional message associated with the event.
- */
 class Event {
 protected:
-    long time; /**< Time at which the event occurs. */
-    EventType event_type; /**< Type of the event. */
-    std::string message; /**< Message associated with the event. */
+    double time;
+    EventType event_type;
+    std::string message;
 public:
-    /**
-     * @brief Constructs an Event object with the provided time.
-     * @param time The time at which the event occurs.
-     */
-    explicit Event(long time);
-
-    /**
-     * @brief Virtual destructor for the Event class.
-     */
+    explicit Event(double time);
     virtual ~Event() = 0;
-
-    /**
-     * @brief Retrieves the time at which the event occurs.
-     * @return The time of the event.
-     */
-    [[nodiscard]] long getTime() const;
-
-    /**
-     * @brief Retrieves the type of the event.
-     * @return The type of the event.
-     */
+    [[nodiscard]] double getTime() const;
     [[nodiscard]] EventType getEventType() const;
-
-    /**
-     * @brief Retrieves the message associated with the event.
-     * @return The message associated with the event.
-     */
     std::string getMessage();
 };
